+++ src/main/java/com/ebicep/warlords/pve/bountysystem/bounties/HuntMithra.java
@@ -52,7 +52,7 @@ public class HuntMithra extends AbstractBounty implements TracksDuringGame, Dail
     }
 
     @Override
-    public int getNewValue() {
+    public long getNewValue() {
         return newKills;
     }
