+++ src/main/java/com/ebicep/warlords/pve/bountysystem/trackers/TracksDuringGame.java
@@ -80,6 +80,6 @@ public interface TracksDuringGame {
         bounty.setValue(bounty.getValue() + getNewValue());
     }
 
-    int getNewValue();
+    long getNewValue();
 
 }
